+++ src/common.h
@@ -42,7 +42,7 @@ void *xalloc(size_t num, size_t size, char* msg);
  *  @returns string pointer.
  *
  */
-char *substring(char *string, int position, int length);
+char *substring(char *string, size_t position, size_t length);
 
 /** Globally replace a char in a char*
  *
